grooveDjango/requirements.txt
Django==3.1.7
djangorestframework==3.12.2
django-cors-headers==3.7.0
djoser==2.1.0
pillow==8.1.1
stripe==2.56.0
django-admin-thumbnails==0.2.6
django-mptt==0.13.4
django-js-asset==1.2.2